Refactor watch logic to its own file.
